PR remarks refactoring, tuning example values for offer in Swagger as QA asked
