+++ backend/src/main/java/com/intive/shopme/config/ApiUrl.java
@@ -1,0 +1,9 @@
+package com.intive.shopme.config;
+
+public class ApiUrl {
+    public static final String OFFERS_PATH = "/offers";
+    public static final String CATEGORIES_PATH = "/categories";
+
+    private ApiUrl() {
+    }
+}
