+++ backend/src/main/java/com/intive/shopme/config/AppConfig.java
@@ -1,13 +1,13 @@
 package com.intive.shopme.config;
 
 public class AppConfig {
-    public static final String OFFERS_PATH = "/offers";
-    public static final String CATEGORIES_PATH = "/categories";
-
     public static final int USER_NAME_MIN_LENGTH = 3;
     public static final int USER_NAME_MAX_LENGTH = 20;
 
     public static final int OFFER_TITLE_MAX_LENGTH = 30;
     public static final int OFFER_DESCRIPTION_MAX_LENGTH = 500;
     public static final int USER_DESCRIPTION_MAX_LENGTH = 800;
+
+    private AppConfig() {
+    }
 }
