+++ backend/src/main/java/com/intive/shopme/controller/CategoryController.java
@@ -5,6 +5,7 @@ import com.intive.shopme.service.CategoryService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
+//import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,7 +14,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
 
-import static com.intive.shopme.config.AppConfig.CATEGORIES_PATH;
+import static com.intive.shopme.config.ApiUrl.CATEGORIES_PATH;
 
 @RestController
 @RequestMapping(value = CATEGORIES_PATH)
