+++ backend/src/main/java/com/intive/shopme/controller/filter/OfferSpecification.java
@@ -17,23 +17,22 @@ public class OfferSpecification implements Specification<Offer> {
     @Override
     public Predicate toPredicate(Root<Offer> root, CriteriaQuery<?> query, CriteriaBuilder builder) {
 
-        Predicate result;
+        Predicate result = null;
         switch (criteria.getOperation()) {
-            case ">":
+            case "≥":
                 result = builder.greaterThanOrEqualTo(root.<String>get(criteria.getKey()), criteria.getValue().toString());
                 break;
-            case "<":
+            case "≤":
                 result = builder.lessThanOrEqualTo(root.<String>get(criteria.getKey()), criteria.getValue().toString());
                 break;
             case ":":
                 if (root.get(criteria.getKey()).getJavaType() == String.class) {
-                    result = builder.like(builder.lower(root.<String>get(criteria.getKey())), "%" + criteria.getValue() + "%");
+                    result = builder.like(builder.lower(root.<String>get(criteria.getKey())),
+                            "%" + criteria.getValue() + "%");
                 } else {
                     result = builder.equal(root.get(criteria.getKey()), criteria.getValue());
                 }
                 break;
-            default:
-                result = null;
         }
 
         return result;
