+++ backend/src/main/java/com/intive/shopme/model/Category.java
@@ -12,7 +12,7 @@ import javax.persistence.Entity;
 @Data
 public class Category extends Identifiable {
 
-    @ApiModelProperty(value = "Represents category's name", required = true, example = "nowa kategoria", position = 2)
+    @ApiModelProperty(value = "Represents category's name", required = true, position = 2, example = "inne")
     private String name;
 
 }
